# openapi.yaml
openapi: 3.0.0
info:
  title: Canvus API
  version: 1.2.0
  description: |
    The Canvus API provides programmatic access to manage canvases, widgets, users, and system
    resources on a Canvus Server. This specification documents all endpoints available through
    the Canvus Go SDK.

    ## Authentication

    All API requests require authentication using one of the following methods:

    - **API Key**: Pass your API key in the `Private-Token` header
    - **Session Token**: After login, use the returned token in the `Private-Token` header

    ## Base URL

    The API base URL follows the format: `https://{server}/api/v1/`

    ## Rate Limiting

    The API implements rate limiting. When rate limited, requests return HTTP 429.
    The SDK includes automatic retry with exponential backoff.

    ## Error Handling

    Errors are returned as JSON objects with `code` and `message` fields.
    HTTP status codes indicate the error category (4xx for client errors, 5xx for server errors).
  contact:
    name: Canvus SDK Support
    url: https://github.com/jaypaulb/Canvus-Go-API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://{server}/api/v1
    description: Canvus Server API
    variables:
      server:
        default: localhost
        description: Canvus server hostname or IP address
...
security:
  - ApiKeyAuth: []
paths:
...
  /canvases:
    get:
      tags:
        - Canvases
      summary: List all canvases
      description: Retrieves all canvases accessible to the authenticated user.
      operationId: listCanvases
      responses:
        '200':
          description: List of canvases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Canvas'
              example:
                - id: "canvas-123"
                  name: Project Canvas
                  access: owner
                  asset_size: 1024000
                  created_at: "2024-01-01T00:00:00Z"
                  folder_id: "folder-1"
                  in_trash: false
                  mode: normal
                  modified_at: "2024-01-15T10:00:00Z"
                  preview_hash: "abc123"
                  state: active
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Canvases
      summary: Create canvas
      description: Creates a new canvas.
      operationId: createCanvas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCanvasRequest'
            example:
              name: New Canvas
              folder_id: "folder-1"
      responses:
        '201':
          description: Canvas created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Canvas'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Private-Token
      description: API key for authentication. Obtain from user access tokens.
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: bad_request
            message: Invalid request body
            request_id: req-xyz789
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: unauthorized
            message: Invalid or missing API key
            request_id: req-abc123
...
  schemas:
    APIError:
      type: object
      description: Standard error response returned by the API
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        request_id:
          type: string
          description: Unique request ID for debugging and support
      example:
        code: not_found
        message: Resource not found
        request_id: req-abc123
...
    Canvas:
      type: object
      description: Canvas resource containing widgets and content
      properties:
        id:
          type: string
          description: Unique canvas identifier
        name:
          type: string
          description: Canvas name
        access:
          type: string
          description: User's access level (owner, editor, viewer)
        asset_size:
          type: integer
          format: int64
          description: Total size of assets in bytes
        created_at:
          type: string
          format: date-time
          description: Canvas creation timestamp
        folder_id:
          type: string
          description: ID of the parent folder
        in_trash:
          type: boolean
          description: Whether the canvas is in trash
        mode:
          type: string
          enum: [normal, demo]
          description: Canvas mode (normal or demo)
        modified_at:
          type: string
          format: date-time
          description: Last modification timestamp
        preview_hash:
          type: string
          description: Hash for the canvas preview image
        state:
          type: string
          description: Canvas state (e.g., active, archived)
    CreateCanvasRequest:
      type: object
      description: Payload for creating a new canvas
      properties:
        name:
          type: string
          description: Canvas name
        folder_id:
          type: string
          description: Parent folder ID (optional)
...